Regenerate scalar and turn-event reports after resolving the integration conflicts, and refresh the reccmp progress baselines on the merged tree.\n\nVerification: just precommit; just format-check src/game/military/TArmyMgr.cpp src/game/military/TArmyStack.cpp. Runtime PR suite: 5 passed.

+++ include/game/military/TMilitaryUnit.h
@@ -65,7 +65,7 @@ public:
   // --- TObject/TUnit overrides ---
   void ReadFrom(TStream* stream) override;
   void WriteTo(TStream* stream) override;
-  void MoveTo(int nTileIndex) override;
+  void MoveTo(short nTileIndex) override;
   void DetachUnitOrderFromOwnerAndReset() override;
 
   // --- TMilitaryUnit virtual functions ---
